Move `MoveOrder` from game to move module, improve absolute imports
