+++ src/core/index.ts
@@ -2,5 +2,5 @@ export { Game } from "../game.js";
 export { Board, CaptureOptions } from "./Board.js";
 export { CheckManager } from "./CheckManager.js";
 export { PromotionManager } from "./PromotionManager.js";
-export { MoveType, Move } from "./move.js";
+export { MoveType, Move, MoveOrder } from "./move.js";
 export { Tile, MovesPredictionsType } from "./Tile.js";
