+++ tsconfig.json
@@ -2,7 +2,7 @@
   "compilerOptions": {
     "target": "es6",
     "strict": true,
-    "baseUrl": "src",
+    "baseUrl": "./src",
     "outDir": "build",
     "module": "nodenext",
     "moduleResolution": "nodenext",
@@ -10,15 +10,15 @@
     "resolveJsonModule": true,
     "esModuleInterop": true,
     "paths": {
-      "@lc/core":            [ "core/index.ts" ],
-      "@lc/utils":           [ "utils/index.ts" ],
-      "@lc/geometry":        [ "geometry/index.ts" ],
-      "@lc/pieces":          [ "pieces/index.ts" ],
-      "@lc/piece-movements": [ "pieces/movements/index.ts" ]
+      "@lc/core":            [ "./core/index.ts" ],
+      "@lc/utils":           [ "./utils/index.ts" ],
+      "@lc/geometry":        [ "./geometry/index.ts" ],
+      "@lc/pieces":          [ "./pieces/index.ts" ],
+      "@lc/piece-movements": [ "./pieces/movements/index.ts" ]
     }
   },
   "include": [
-    "src/**/*"
+    "src/**/*",
   ],
   "exclude": [
     "node_modules/**/*"
